v1.0 (#8)
* Create nrchkb-ffmpeg-build-alpine.sh

* Update README.md

* Update README.md

* Update README.md

* Update README.md

* Update README.md

* Create build-bootstrap.sh

* Rename nrchkb-ffmpeg-build-alpine.sh to build-alpine.sh

* Rename nrchkb-ffmpeg-build.sh to build-ubuntu.sh

* Update README.md

* Update README.md

* Update README.md

* Update build-bootstrap.sh

* Update build-bootstrap.sh

* Update build-bootstrap.sh

* Update build-bootstrap.sh

* Update build-bootstrap.sh

* Update build-bootstrap.sh

* Update build-bootstrap.sh

* Args

* Delete .DS_Store

* Create build-alpine.sh

* Update build-bootstrap.sh

* Feature/pass args (#7)

* Pass all args deeper

* Display OS in header

* GH Actions aligned to args

* check package managers

* Update build-bootstrap.sh

* Update build-bootstrap.sh

* Update build-bootstrap.sh

* Use string length

* Update build-bootstrap.sh

* Update README.md

+++ build-debian.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-# nrchkb-ffmpeg-build Version 0.9
+# nrchkb-ffmpeg-build Version 1.0
 
 # MIT License
 
@@ -30,14 +30,28 @@ Red=$'\e[0;31m'
 Yellow=$'\e[1;33m'
 End=$'\e[0m'
 
+# Defaults
+JOBS=3           # Jobs Value
+JOBSPARAM=false  # Arg provided
+OMX="n"          # OMX Choice Value
+OMXPARAM=false   # Arg provided
+FDK="y"          # FDK Choice Value
+FDKPARAM=false   # Arg provided
+FLAGSYN="n"      # Flags Choice Value
+FLAGS=""         # Flasg Value
+FLAGSPARAM=false # Arg provided
+MODE=0           # Mode Value
+MODEPARAM=false  # Arg provided
+INTERACTIVE="y"  # Interactive
+
 # Print Header
 printHeader() {
 
     printf "\033c"
     echo
     echo " ---------------------------------------------------------"
-    echo " |                                                       |"
-    echo " |               P&M FFmpeg Build Script                 |"
+    echo " |                                                   1.0 |"
+    echo " |          P&M FFmpeg Build Script (Debian)             |"
     echo " |   An FFmpeg build & installation utility for NRCHKB   |"
     echo " |                                                       |"
     echo " ---------------------------------------------------------"
@@ -65,17 +79,18 @@ menu() {
     echo "   If you have previously run this script, running it again will update your software."
     echo
     printf "   Choice: "
-    read -r Mode
-
-    if [[ "$Mode" = "q" ]]; then
+    read -r
+    if [[ "$REPLY" = "q" ]]; then
         exit 0
     fi
 
-    if [[ $Mode -gt 5 || $Mode -lt 1 ]]; then
+    MODE=$REPLY
+
+    if [[ $MODE -gt 5 || $MODE -lt 1 ]]; then
         printHeader
         menu
     fi
-    processOptions "$Mode"
+    processOptions $MODE
 }
 
 # Error Check
@@ -106,7 +121,7 @@ installDependencies() {
     echo
     sudo apt install -y pkg-config autoconf automake libtool git wget make g++ gcc nasm yasm
 
-    LibXCheck=(sudo apt info libx264-dev)
+    sudo apt info libx264-dev
 
     if [[ $? -gt 0 ]]; then
         installLibx264
@@ -118,7 +133,10 @@ installDependencies() {
 
 # Install Libx264
 installLibx264() {
-    cd ~ || { echo "cd failed, aborting at installLibx264:01"; exit 1; }
+    cd ~ || {
+        echo "cd failed, aborting at installLibx264:01"
+        exit 1
+    }
     echo
     echo " ---------------------------------------------------------"
     echo " |                                                       |"
@@ -129,20 +147,29 @@ installLibx264() {
     sudo apt remove -y libx264-dev
     sudo apt purge -y libx264-dev
     git clone https://code.videolan.org/videolan/x264.git
-    cd x264 || { echo "cd failed, aborting at installLibx264:02"; exit 1; }
+    cd x264 || {
+        echo "cd failed, aborting at installLibx264:02"
+        exit 1
+    }
     sudo ./configure --prefix="/usr" --enable-static --enable-pic
     checkForError
-    sudo make -j"$Jobs"
+    sudo make -j"$JOBS"
     checkForError
     sudo make install
     checkForError
     sudo ldconfig
-    cd ~ || { echo "cd failed, aborting at installLibx264:03"; exit 1; }
+    cd ~ || {
+        echo "cd failed, aborting at installLibx264:03"
+        exit 1
+    }
 }
 
 # Install Libfdk
 installLibfdk() {
-    cd ~ || { echo "cd failed, aborting at installLibfdk:01"; exit 1; }
+    cd ~ || {
+        echo "cd failed, aborting at installLibfdk:01"
+        exit 1
+    }
     echo
     echo " ---------------------------------------------------------"
     echo " |                                                       |"
@@ -150,24 +177,40 @@ installLibfdk() {
     echo " |                                                       |"
     echo " ---------------------------------------------------------"
     echo
+
+    sudo apt info libfdk-aac-dev
+    if [[ $? = 0 ]]; then
+        sudo apt install -y libfdk-aac-dev
+        return
+    fi
+
     sudo apt remove -y libfdk-aac-dev
     sudo apt purge -y libfdk-aac-dev
     git clone https://github.com/mstorsjo/fdk-aac.git
-    cd fdk-aac || { echo "cd failed, aborting at installLibfdk:02"; exit 1; }
+    cd fdk-aac || {
+        echo "cd failed, aborting at installLibfdk:02"
+        exit 1
+    }
     sudo ./autogen.sh
     sudo ./configure --prefix="/usr" --enable-static --disable-shared
     checkForError
-    sudo make -j"$Jobs"
+    sudo make -j"$JOBS"
     checkForError
     sudo make install
     checkForError
     sudo ldconfig
-    cd ~ || { echo "cd failed, aborting at installLibfdk:03"; exit 1; }
+    cd ~ || {
+        echo "cd failed, aborting at installLibfdk:03"
+        exit 1
+    }
 }
 
 # Install FFmpeg
 installFFmpeg() {
-    cd ~ || { echo "cd failed, aborting at installFFmpeg:01"; exit 1; }
+    cd ~ || {
+        echo "cd failed, aborting at installFFmpeg:01"
+        exit 1
+    }
     echo
     echo " ---------------------------------------------------------"
     echo " |                                                       |"
@@ -180,11 +223,14 @@ installFFmpeg() {
     wget -O ffmpeg-snapshot.tar.bz2 https://ffmpeg.org/releases/ffmpeg-snapshot.tar.bz2
     echo "Extracting source code..."
     tar xjf ffmpeg-snapshot.tar.bz2
-    cd ffmpeg || { echo "cd failed, aborting at installFFmpeg:02"; exit 1; }
+    cd ffmpeg || {
+        echo "cd failed, aborting at installFFmpeg:02"
+        exit 1
+    }
 
     CMD="--prefix=\"/usr\" --enable-nonfree --enable-gpl --enable-hardcoded-tables --disable-ffprobe --disable-ffplay --enable-libx264"
 
-    if [[ "$FDK" != "n" ]]; then
+    if [[ "$FDK" = "y" ]]; then
         CMD="$CMD --enable-libfdk-aac"
     fi
 
@@ -200,11 +246,14 @@ installFFmpeg() {
 
     sudo ./configure $CMD
     checkForError
-    sudo make -j"$Jobs"
+    sudo make -j"$JOBS"
     checkForError
     sudo make install
     checkForError
-    cd ~ || { echo "cd failed, aborting at installFFmpeg:03"; exit 1; }
+    cd ~ || {
+        echo "cd failed, aborting at installFFmpeg:03"
+        exit 1
+    }
 }
 
 # Clear Up
@@ -217,53 +266,86 @@ cleanDirectory() {
 
 # Ask for Threads
 getJobsCount() {
+
+    if [[ $JOBSPARAM = true || "$INTERACTIVE" = "n" ]]; then
+        return
+    fi
+
     echo
     echo "   ${Yellow}How many simultaneous jobs would you like to use for build processes (if needed)${End}"
     echo
     echo "   The more you specify - the higher chance of CPU throttling and memory constraints"
     printf "   we recommend no more than 3 for a Pi 4 (1-4): "
-    read -r Jobs
-    if [[ $Jobs != 1 && $Jobs != 2 && $Jobs != 4 ]]; then
-        Jobs=3
+    read -r
+    if [[ $REPLY -lt 5 && $REPLY -gt 0 ]]; then
+        JOBS=$REPLY
     fi
 }
 
 # Ask for omx
 getOMX() {
+
+    if [[ $OMXPARAM = true || "$INTERACTIVE" = "n" ]]; then
+        return
+    fi
+
     echo
     echo "   ${Yellow}Would you like to enable 'h264_omx'?${End}"
     echo
     echo "   Note: 'h264_omx' is deprecated and should not be used on new installs."
     printf "   If you already use it, choose yes here. Enter (y/n): "
-    read -r OMX
+    read -r
+    if [[ "$REPLY" = "y" || "$REPLY" = "n" ]]; then
+        OMX="$REPLY"
+    fi
 }
 
 # Ask for FDK
 getFDK() {
+
+    if [[ $FDKPARAM = true || "$INTERACTIVE" = "n" ]]; then
+        return
+    fi
+
     echo
     echo "   ${Yellow}Would you like to enable 'libfdk-aac'?${End}"
     echo
     echo "   Note: 'libfdk-aac' is needed for HomeKit audio. We recommend enabling libfdk-aac."
     printf "   If you are running Option 4, you can enable this lib. Enter (y/n): "
-    read -r FDK
+    read -r
+    if [[ "$REPLY" = "y" || "$REPLY" = "n" ]]; then
+        FDK="$REPLY"
+    fi
 }
 
 # Get Compile Flags
 getFlags() {
+
+    if [[ $FLAGSPARAM = true || "$INTERACTIVE" = "n" ]]; then
+        return
+    fi
 
     echo
     echo "   ${Yellow}Would you like to add any extra FFmpeg compile flags?"
     echo
     echo "   ADVANCED: ${End}Compile flags could be added to enable libx265 or the countless others"
     printf "   You are responsible for ensuring any required parts are installed (y/n): "
-    read -r FLAGSYN
+    read -r
+    if [[ "$REPLY" = "y" || "$REPLY" = "n" ]]; then
+        FLAGSYN="$REPLY"
+        if [[ "$FLAGSYN" = "y" ]]; then
+            echo
+            echo "   ${Yellow}Please enter your compile flags below, separated by a space${End}"
+            echo
+            printf "   Example '--enable-libx265 --enable-libopus' : "
+            read -r
+            if [[ ${#REPLY} -gt 0 ]]; then
+                FLAGS="$REPLY"
+            else
+                FLAGSYN="n"
+            fi
 
-    if [[ "$FLAGSYN" = "y" ]]; then
-        echo
-        echo "   ${Yellow}Please enter your compile flags below, separated by a space${End}"
-        echo
-        printf "   Example '--enable-libx265 --enable-libopus' : "
-        read -r FLAGS
+        fi
     fi
 
 }
@@ -278,8 +360,8 @@ stopWatch() {
         echo "   Start time: ${startTime}"
         echo "   End time:   ${endTime}"
         echo "   Duration:   ${durationEpoch} seconds"
-        echo "   Max jobs:   $Jobs"
-        echo "   Option:     $Mode"
+        echo "   Max jobs:   $JOBS"
+        echo "   Option:     $MODE"
         echo
     else
         startEpoch=$(date +%s)
@@ -297,10 +379,12 @@ processOptions() {
         stopWatch "start"
         installDependencies
         stopWatch "stop"
-        echo "   ${Yellow}All Done!${End} ...press enter"
-        read -r
-        printHeader
-        menu
+        if [ "$INTERACTIVE" = "y" ]; then
+            echo "   ${Yellow}All Done!${End} ...press enter"
+            read -r
+            printHeader
+            menu
+        fi
         ;;
 
     2)
@@ -308,10 +392,12 @@ processOptions() {
         stopWatch "start"
         installLibfdk
         stopWatch "stop"
-        echo "   ${Yellow}All Done!${End} ...press enter"
-        read -r
-        printHeader
-        menu
+        if [ "$INTERACTIVE" = "y" ]; then
+            echo "   ${Yellow}All Done!${End} ...press enter"
+            read -r
+            printHeader
+            menu
+        fi
         ;;
 
     3)
@@ -322,10 +408,12 @@ processOptions() {
         stopWatch "start"
         installFFmpeg
         stopWatch "stop"
-        echo "   ${Yellow}All Done!${End} ...press enter"
-        read -r
-        printHeader
-        menu
+        if [ "$INTERACTIVE" = "y" ]; then
+            echo "   ${Yellow}All Done!${End} ...press enter"
+            read -r
+            printHeader
+            menu
+        fi
         ;;
 
     4)
@@ -336,31 +424,91 @@ processOptions() {
         getFlags
         stopWatch "start"
         installDependencies
-        if [[ "$FDK" != "n" ]]; then
+        if [[ "$FDK" = "y" ]]; then
             installLibfdk
         fi
         installFFmpeg
         cleanDirectory
         stopWatch "stop"
-        echo "   ${Yellow}All Done!${End} ...press enter"
-        read -r
-        printHeader
-        menu
+        if [ "$INTERACTIVE" = "y" ]; then
+            echo "   ${Yellow}All Done!${End} ...press enter"
+            read -r
+            printHeader
+            menu
+        fi
         ;;
 
     5)
         stopWatch "start"
         cleanDirectory
         stopWatch "stop"
-        echo "   ${Yellow}All Done!${End} ...press enter"
-        read -r
-        printHeader
-        menu
+        if [ "$INTERACTIVE" = "y" ]; then
+            echo "   ${Yellow}All Done!${End} ...press enter"
+            read -r
+            printHeader
+            menu
+        fi
         ;;
     esac
 }
 
 # Entry Point
-cd ~ || { echo "cd failed, aborting"; exit 1; }
-printHeader
-menu
+cd ~ || {
+    echo "cd failed, aborting"
+    exit 1
+}
+
+while [ $# -gt 0 ]; do
+
+    case "$1" in
+
+    --help | -h)
+        echo "Options:"
+        echo
+        echo " --interactive -i    [y|n]             Default: y        Disables interactive mode"
+        echo " --mode        -m    [1|2|3|4|5]                         Sets the mode (requires --interactive 0, ignored otherwise)"
+        echo " --max-jobs    -j    [#]               Default: 3        Set maximum number of builds jobs"
+        echo " --libfdk-aac  -a    [y|n]             Default: y        Compile/enable libfdk-aac"
+        echo " --h264-omx    -x    [y|n]             Default: n        Enable h264-omx"
+        echo " --extra-flags -f    [\"--*-* --*-*\"]   Default: empty    Provides compile flags for ffmpeg"
+        echo
+        echo " Example: ./build-debian.sh --interactive n --max-jobs 2 --mode 4"
+        echo
+        exit 0
+        ;;
+    --interactive | -i)
+        INTERACTIVE="$2"
+        ;;
+    --mode | -m)
+        MODE=$2
+        MODEPARAM=true
+        ;;
+    --max-jobs | -j)
+        JOBS=$2
+        JOBSPARAM=true
+        ;;
+    --h264-omx | -x)
+        OMX="$2"
+        OMXPARAM=true
+        ;;
+    --libfdk-aac | -a)
+        FDK="$2"
+        FDKPARAM=true
+        ;;
+    --extra-flags | -f)
+        FLAGSYN="y"
+        FLAGS="$2"
+        FLAGSPARAM=true
+        ;;
+    esac
+    shift
+done
+
+if [ "$INTERACTIVE" = "n" ]; then
+    printHeader
+    processOptions $MODE
+
+else
+    printHeader
+    menu
+fi
